.idea/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
msal==1.28.0
atlassian-python-api==3.41.11
apscheduler==3.10.4
python-multipart==0.0.9

# ── SharePoint file extraction ─────────────────────────────────────────────
python-docx==1.1.0       # Word documents (.docx)
python-pptx==0.6.23      # PowerPoint (.pptx)
openpyxl==3.1.2          # Excel spreadsheets (.xlsx)
pdfplumber==0.11.0       # PDF text extraction

# SharePoint Online auth (no Azure AD needed)
shareplum==0.5.1         # handles Office365/NTLM auth + SharePoint REST
requests==2.31.0         # HTTP client
requests-ntlm==1.3.0    # NTLM fallback for on-premise
